requirements.txt
flask==3.1.3
flask-cors==6.0.2
mysql-connector-python==9.6.0
pandas==2.2.2
matplotlib==3.9.0
seaborn==0.13.2
python-dotenv==1.0.1